movie_ideas/serializers.py
# każdy model ma osobny serializer

from .models import Idea, Vote
from rest_framework import serializers

class IdeaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Idea
        fields = '__all__'

class VoteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vote
        fields = '__all__'